Reference

FastMultipole.DerivativesSwitch

FastMultipole.DerivativesSwitch

FastMultipole.DerivativesSwitch

FastMultipole.DerivativesSwitch

FastMultipole.MultiTree

FastMultipole.ProbeSystem

FastMultipole.ProbeSystem

FastMultipole.ProbeSystem

FastMultipole.SortWrapper

FastMultipole.Tree

FastMultipole.EmptyTree

FastMultipole.L2B_velocity

FastMultipole.L2B_velocity_gradient

FastMultipole.add_line!

FastMultipole.d2rdx2

FastMultipole.direct!

FastMultipole.direct!

FastMultipole.drdx

FastMultipole.fmm!

FastMultipole.fmm!

FastMultipole.fmm!

FastMultipole.fmm!

FastMultipole.fmm!

FastMultipole.fmm!

FastMultipole.populate_influence_matrix!

FastMultipole.populate_influence_matrix!

FastMultipole.populate_influence_matrix!

FastMultipole.populate_influence_matrix!

FastMultipole.populate_influence_matrix!

FastMultipole.reset!

FastMultipole.resort!

FastMultipole.unsort!

FastMultipole.DerivativesSwitch — Type

DerivativesSwitch

Switch indicating whether the scalar potential, vector potential, velocity, and/or velocity gradient should be computed for a target system. Information is stored as type parameters, allowing the compiler to compile away if statements.

FastMultipole.DerivativesSwitch — Method

DerivativesSwitch(scalar_potential, vector_potential, velocity, velocity_gradient)

Constructs a tuple of DerivativesSwitch objects.

Arguments

scalar_potential::Vector{Bool}: a vector of ::Bool indicating whether the scalar potential should be computed for each target system

vector_potential::Vector{Bool}: a vector of ::Bool indicating whether the vector potential should be computed for each target system

velocity::Vector{Bool}: a vector of ::Bool indicating whether the velocity should be computed for each target system

velocity_gradient::Vector{Bool}: a vector of ::Bool indicating whether the velocity gradient should be computed for each target system

FastMultipole.DerivativesSwitch — Method

DerivativesSwitch(scalar_potential, vector_potential, velocity, velocity_gradient)

Constructs a single DerivativesSwitch object.

Arguments

scalar_potential::Bool: a ::Bool indicating whether the scalar potential should be computed for the target system

vector_potential::Bool: a ::Bool indicating whether the vector potential should be computed for the target system

velocity::Bool: a ::Bool indicating whether the velocity should be computed for the target system

velocity_gradient::Bool: a ::Bool indicating whether the velocity gradient should be computed for the target system

FastMultipole.DerivativesSwitch — Method

DerivativesSwitch(scalar_potential, vector_potential, velocity, velocity_gradient, target_systems)

Constructs a ::Tuple of indentical DerivativesSwitch objects of the same length as target_systems (if it is a ::Tuple), or a single DerivativesSwitch (if target_system is not a ::Tuple)

Arguments

scalar_potential::Bool: a ::Bool indicating whether the scalar potential should be computed for each target system

vector_potential::Bool: a ::Bool indicating whether the vector potential should be computed for each target system

velocity::Bool: a ::Bool indicating whether the velocity should be computed for each target system

velocity_gradient::Bool: a ::Bool indicating whether the velocity gradient should be computed for each target system

FastMultipole.MultiTree — Type

bodies[indexlist] is the same sort operation as performed by the tree sortedbodies[inverseindexlist] undoes the sort operation performed by the tree

FastMultipole.ProbeSystem — Type

ProbeSystem

Convenience system for defining locations at which the potential, velocity, or velocity gradient may be desired.

FastMultipole.ProbeSystem — Method

ProbeSystem(positions; kwargs...)

A convenience system constructor for calculating the influence of source systems at locations not already described by a system object. It behaves like a system whose elements induce a null potential field.

Arguments

positions::Vector{SVector{3,Float64}}: a vector of position vectors of each probe

Optional Arguments

scalar_potential::Bool: whether or not to compute the scalar potential at each probe location

vector_potential::Bool: whether or not to compute the vector potential at each probe location

velocity::Bool: whether or not to compute the velocity at each probe location

velocity_gradient::Bool: whether or not to compute the velocity gradient at each probe location

FastMultipole.ProbeSystem — Method

ProbeSystem(positions; kwargs...)

Dispatch of ProbeSystem accepting a matrix of horizontally concatenated column vectors describing the position of each probe. Optional arguments are identical.

FastMultipole.SortWrapper — Method

SortWrapper(system)

Convenience wrapper for systems whose elements cannot be sorted in-place (e.g. structured grids). The resulting object is treated like any other system.

FastMultipole.Tree — Type

abstract type Tree{TF,P} end

Supertype of all octree structures with TF the floating point type and P the expansion order.

FastMultipole.EmptyTree — Method

EmptyTree(system)

Returns an empty tree. Used if system is empty.

Arguments

system: the system from which a tree is to be created

Returns

tree: if typeof(system)<:Tuple, a ::MultiTree is returned; otherwise, a ::SingleTree is returned

FastMultipole.L2B_velocity — Method

Does not include the rotation matrix R yet. Eq 28 in S&L

FastMultipole.L2B_velocity_gradient — Method

Eq 35-37 S&L

FastMultipole.add_line! — Method

add_line!(probes::ProbeSystem, x1, x2, n_probes, i_last)

Adds n_probes probes in a line between x1 and x2. Specifically, they are added at the midpoint of equally partiti oned segments of the line.

Arguments

probes::ProbeSystem: the probe system whose .position field is to be updated

x1::Vector{Float64}: the first point defining the line along which probes are to be added

x2::Vector{Float64}: the second point defining the line along which probes are to be added

n_probes::Int: number of probes to be added

i_last::Int: index of the last probe in probes

Output

i_last::Int: updated index of the last probe in probes

FastMultipole.d2rdx2 — Method

drk/dxidx_j

FastMultipole.direct! — Method

direct!(target_system, source_system; derivatives_switches)

Applies all interactions of source_system acting on target_system without multipole acceleration.

Arguments

target_system: either

a system object for which compatibility functions have been overloaded, or

a tuple of system objects for which compatibility functions have been overloaded

source_system: either

a system object for which compatibility functions have been overloaded, or

a tuple of system objects for which compatibility functions have been overloaded

Optional Arguments

scalar_potential::Bool: either a ::Bool or a ::AbstractVector{Bool} of length length(target_systems) indicating whether each system should receive a scalar potential from source_systems

vector_potential::Bool: either a ::Bool or a ::AbstractVector{Bool} of length length(target_systems) indicating whether each system should receive a vector potential from source_systems

velocity::Bool: either a ::Bool or a ::AbstractVector{Bool} of length length(target_systems) indicating whether each system should receive a velocity from source_systems

velocity_gradient::Bool: either a ::Bool or a ::AbstractVector{Bool} of length length(target_systems) indicating whether each system should receive a velocity gradient from source_systems

FastMultipole.direct! — Method

direct!(systems; derivatives_switches)

Applies all interactions of systems acting on itself without multipole acceleration.

Arguments

systems: either

a system object for which compatibility functions have been overloaded, or

a tuple of system objects for which compatibility functions have been overloaded

Optional Arguments

scalar_potential::Bool: either a ::Bool or a ::AbstractVector{Bool} of length length(target_systems) indicating whether each system should receive a scalar potential from source_systems

vector_potential::Bool: either a ::Bool or a ::AbstractVector{Bool} of length length(target_systems) indicating whether each system should receive a vector potential from source_systems

velocity::Bool: either a ::Bool or a ::AbstractVector{Bool} of length length(target_systems) indicating whether each system should receive a velocity from source_systems

velocity_gradient::Bool: either a ::Bool or a ::AbstractVector{Bool} of length length(target_systems) indicating whether each system should receive a velocity gradient from source_systems

FastMultipole.drdx — Method

drj/dxi

FastMultipole.fmm! — Method

fmm!(target_systems, source_systems; kwargs...)

Apply all interactions of source_systems acting on target_systems using the fast multipole method. Assumes compatibility functions have been overloaded for both source and target systems.

Arguments

target_systems: either

a system object for which compatibility functions have been overloaded, or

a tuple of system objects for which compatibility functions have been overloaded

source_systems: either

a system object for which compatibility functions have been overloaded, or

a tuple of system objects for which compatibility functions have been overloaded

Optional Arguments

expansion_order::Int: the expansion order to be used

leaf_size_source::Int: maximum number of bodies from source_systems allowed in a leaf-level branch

leaf_size_target::Int: maximum number of bodies from target_systems allowed in a leaf-level branch

multipole_threshold::Float64: number between 0 and 1 (often denoted theta in [0,1]) controls the accuracy by determining the non-dimensional distance after which multipoles are used; 0 means an infinite distance (no error, high cost), and 1 means barely convergent (high error, low cost)

scalar_potential::Bool: either a ::Bool or a ::AbstractVector{Bool} of length length(target_systems) indicating whether each system should receive a scalar potential from source_systems

vector_potential::Bool: either a ::Bool or a ::AbstractVector{Bool} of length length(target_systems) indicating whether each system should receive a vector potential from source_systems

velocity::Bool: either a ::Bool or a ::AbstractVector{Bool} of length length(target_systems) indicating whether each system should receive a velocity from source_systems

velocity_gradient::Bool: either a ::Bool or a ::AbstractVector{Bool} of length length(target_systems) indicating whether each system should receive a velocity gradient from source_systems

upward_pass::Bool: whether or not to form the multipole expansions from source bodies and translate them upward in the source tree

horizontal_pass::Bool: whether or not to transform multipole expansions from the source tree into local expansions in the target tree

downward_pass::Bool: whether or not to translate local expansions down to the leaf level of the target tree and evaluate them

nearfield::Bool: indicates whether near-field (comuted without multipoles) interactions should be included

farfield::Bool: indicates whether far-field (comuted with multipoles) interactions should be included

self_induced::Bool: indicates whether to include the interactions of each leaf-level branch on itself

unsort_source_bodies::Bool: indicates whether or not to undo the sort operation used to generate the octree for the source_systems

unsort_target_bodies::Bool: indicates whether or not to undo the sort operation used to generate the octree for the target_systems

source_shink_recenter::Bool: indicates whether or not to resize branches for the source_systems octree after it is created to increase computational efficiency

target_shink_recenter::Bool: indicates whether or not to resize branches for the target_systems octree after it is created to increase computational efficiency

save_tree_source::Bool: indicates whether or not to save a VTK file for visualizing the source octree

save_tree_target::Bool: indicates whether or not to save a VTK file for visualizing the target octree

save_name_source::String: name and path of the source octree visualization if save_tree == true

save_name_target::String: name and path of the target octree visualization if save_tree == true

FastMultipole.fmm! — Method

fmm!(systems; kwargs...)

Apply all interactions of systems acting on itself using the fast multipole method. Assumes compatibility functions have been overloaded for both source and target systems.

Arguments

systems: either

a system object for which compatibility functions have been overloaded, or

a tuple of system objects for which compatibility functions have been overloaded

Optional Arguments

expansion_order::Int: the expansion order to be used

leaf_size::Int: maximum number of bodies from systems allowed in a leaf-level branch

multipole_threshold::Float64: number between 0 and 1 (often denoted theta in [0,1]) controls the accuracy by determining the non-dimensional distance after which multipoles are used; 0 means an infinite distance (no error, high cost), and 1 means barely convergent (high error, low cost)

scalar_potential::Bool: either a ::Bool or a ::AbstractVector{Bool} of length length(systems) indicating whether each system should receive a scalar potential from source_systems

vector_potential::Bool: either a ::Bool or a ::AbstractVector{Bool} of length length(systems) indicating whether each system should receive a vector potential from source_systems

velocity::Bool: either a ::Bool or a ::AbstractVector{Bool} of length length(systems) indicating whether each system should receive a velocity from source_systems

velocity_gradient::Bool: either a ::Bool or a ::AbstractVector{Bool} of length length(systems) indicating whether each system should receive a velocity gradient from source_systems

upward_pass::Bool: whether or not to form the multipole expansions from source bodies and translate them upward in the source tree

horizontal_pass::Bool: whether or not to transform multipole expansions from the source tree into local expansions in the target tree

downward_pass::Bool: whether or not to translate local expansions down to the leaf level of the target tree and evaluate them

nearfield::Bool: indicates whether near-field (comuted without multipoles) interactions should be included in the direct_list

farfield::Bool: indicates whether far-field (comuted with multipoles) interactions should be included in the m2l_list

self_induced::Bool: indicates whether to include the interactions of each leaf-level branch on itself in the direct_list

unsort_bodies::Bool: indicates whether or not to undo the sort operation used to generate the octree for systems

shink_recenter::Bool: indicates whether or not to resize branches for the octree after it is created to increase computational efficiency

save_tree::Bool: indicates whether or not to save a VTK file for visualizing the octree

save_name::String: name and path of the octree visualization if save_tree == true

gpu::Bool: indicates whether or not GPU is to be used for direct interactions

FastMultipole.fmm! — Method

fmm!(target_tree, target_systems, source_tree, source_systems; kwargs...)

Dispatches fmm! using existing ::Tree objects.

Arguments

target_tree::Tree: a <:Tree object (see Tree)

target_systems: either

a system object for which compatibility functions have been overloaded, or

a tuple of system objects for which compatibility functions have been overloaded

source_tree::Tree: a <:Tree object (see Tree)

source_systems: either

a system object for which compatibility functions have been overloaded, or

a tuple of system objects for which compatibility functions have been overloaded

Optional Arguments

multipole_threshold::Float64: number between 0 and 1 (often denoted theta in [0,1]) controls the accuracy by determining the non-dimensional distance after which multipoles are used; 0 means an infinite distance (no error, high cost), and 1 means barely convergent (high error, low cost)

scalar_potential::Bool: either a ::Bool or a ::AbstractVector{Bool} of length length(target_systems) indicating whether each system should receive a scalar potential from source_systems

vector_potential::Bool: either a ::Bool or a ::AbstractVector{Bool} of length length(target_systems) indicating whether each system should receive a vector potential from source_systems

velocity::Bool: either a ::Bool or a ::AbstractVector{Bool} of length length(target_systems) indicating whether each system should receive a velocity from source_systems

velocity_gradient::Bool: either a ::Bool or a ::AbstractVector{Bool} of length length(target_systems) indicating whether each system should receive a velocity gradient from source_systems

reset_source_tree::Bool: either a ::Bool or a ::AbstractVector{Bool} of length length(source_systems) indicating whether or not to reset the expansions of each source tree

reset_target_tree::Bool: either a ::Bool or a ::AbstractVector{Bool} of length length(source_systems) indicating whether or not to reset the expansions of each target tree

upward_pass::Bool: whether or not to form the multipole expansions from source bodies and translate them upward in the source tree

horizontal_pass::Bool: whether or not to transform multipole expansions from the source tree into local expansions in the target tree

downward_pass::Bool: whether or not to translate local expansions down to the leaf level of the target tree and evaluate them

nearfield::Bool: indicates whether near-field (computed without multipoles) interactions should be included in the direct_list

farfield::Bool: indicates whether far-field (computed with multipoles) interactions should be included in the m2l_list

self_induced::Bool: indicates whether to include the interactions of each leaf-level branch on itself in the direct_list

unsort_source_bodies::Bool: indicates whether or not to undo the sort operation used to generate the octree for source_systems

unsort_target_bodies::Bool: indicates whether or not to undo the sort operation used to generate the octree for target_systems

gpu::Bool: indicates whether or not GPU is to be used for direct interactions

FastMultipole.fmm! — Method

fmm!(target_tree, target_systems, source_tree, source_systems, m2l_list, direct_list, derivatives_switches; kwargs...)

Dispatches fmm! using existing ::Tree objects.

Arguments

target_tree::Tree: a <:Tree object (see Tree)

target_systems: either

a system object for which compatibility functions have been overloaded, or

a tuple of system objects for which compatibility functions have been overloaded

source_tree::Tree: a <:Tree object (see Tree)

source_systems: either

a system object for which compatibility functions have been overloaded, or

a tuple of system objects for which compatibility functions have been overloaded

m2l_list::Vector{SVector{2,Int32}}: list of branch index pairs [i_target, i_source] for which multipole expansions of the source branch are to be transformed to local expansions at the target branch

direct_list::Union{Vector{SVector{2,Int32}}, InteractionList}: list of branch index pairs [i_target, i_source] for which interactions are to be evaluted without multipole expansion (i.e., directly); if typeof(direct_list) <: InteractionList, then prepared influence matrices are used rather than computing direct influences on the fly

derivatives_switches::Union{DerivativesSwitch, Tuple{<:DerivativesSwitch,...}}: switch determining which of scalar potential, vector potential, velocity, and/or velocity gradient are to be computed for each target system

Optional Arguments

multipole_threshold::Float64: number between 0 and 1 (often denoted theta in [0,1]) controls the accuracy by determining the non-dimensional distance after which multipoles are used; 0 means an infinite distance (no error, high cost), and 1 means barely convergent (high error, low cost)

scalar_potential::Bool: either a ::Bool or a ::AbstractVector{Bool} of length length(target_systems) indicating whether each system should receive a scalar potential from source_systems

vector_potential::Bool: either a ::Bool or a ::AbstractVector{Bool} of length length(target_systems) indicating whether each system should receive a vector potential from source_systems

velocity::Bool: either a ::Bool or a ::AbstractVector{Bool} of length length(target_systems) indicating whether each system should receive a velocity from source_systems

velocity_gradient::Bool: either a ::Bool or a ::AbstractVector{Bool} of length length(target_systems) indicating whether each system should receive a velocity gradient from source_systems

upward_pass::Bool: whether or not to form the multipole expansions from source bodies and translate them upward in the source tree

horizontal_pass::Bool: whether or not to transform multipole expansions from the source tree into local expansions in the target tree

downward_pass::Bool: whether or not to translate local expansions down to the leaf level of the target tree and evaluate them

nearfield::Bool: indicates whether near-field (comuted without multipoles) interactions should be included

farfield::Bool: indicates whether far-field (comuted with multipoles) interactions should be included

self_induced::Bool: indicates whether to include the interactions of each leaf-level branch on itself

unsort_source_bodies::Bool: indicates whether or not to undo the sort operation used to generate the octree for source_systems

unsort_target_bodies::Bool: indicates whether or not to undo the sort operation used to generate the octree for target_systems

FastMultipole.fmm! — Method

fmm!(tree, systems; kwargs...)

Dispatches fmm! using an existing ::Tree.

Arguments

tree::Tree: a <:Tree object (see Tree)

systems: either

a system object for which compatibility functions have been overloaded, or

a tuple of system objects for which compatibility functions have been overloaded

Optional Arguments

multipole_threshold::Float64: number between 0 and 1 (often denoted theta in [0,1]) controls the accuracy by determining the non-dimensional distance after which multipoles are used; 0 means an infinite distance (no error, high cost), and 1 means barely convergent (high error, low cost)

scalar_potential::Bool: either a ::Bool or a ::AbstractVector{Bool} of length length(systems) indicating whether each system should receive a scalar potential from source_systems

vector_potential::Bool: either a ::Bool or a ::AbstractVector{Bool} of length length(systems) indicating whether each system should receive a vector potential from source_systems

velocity::Bool: either a ::Bool or a ::AbstractVector{Bool} of length length(systems) indicating whether each system should receive a velocity from source_systems

velocity_gradient::Bool: either a ::Bool or a ::AbstractVector{Bool} of length length(systems) indicating whether each system should receive a velocity gradient from source_systems

upward_pass::Bool: whether or not to form the multipole expansions from source bodies and translate them upward in the source tree

horizontal_pass::Bool: whether or not to transform multipole expansions from the source tree into local expansions in the target tree

downward_pass::Bool: whether or not to translate local expansions down to the leaf level of the target tree and evaluate them

nearfield::Bool: indicates whether near-field (computed without multipoles) interactions should be included in the direct_list

farfield::Bool: indicates whether far-field (computed with multipoles) interactions should be included in the m2l_list

self_induced::Bool: indicates whether to include the interactions of each leaf-level branch on itself in the direct_list

unsort_bodies::Bool: indicates whether or not to undo the sort operation used to generate the octree for systems

gpu::Bool: indicates whether or not GPU is to be used for direct interactions

FastMultipole.fmm! — Method

fmm!(tree, systems; kwargs...)

Dispatches fmm! using an existing ::Tree.

Arguments

tree::Tree: a <:Tree object (see Tree)

systems: either

a system object for which compatibility functions have been overloaded, or

a tuple of system objects for which compatibility functions have been overloaded

m2l_list::Vector{SVector{2,Int32}}: list of branch index pairs [i_target, i_source] for which multipole expansions of the source branch are to be transformed to local expansions at the target branch

direct_list::Union{Vector{SVector{2,Int32}}, InteractionList}: list of branch index pairs [i_target, i_source] for which interactions are to be evaluted without multipole expansion (i.e., directly); if typeof(direct_list) <: InteractionList, then prepared influence matrices are used rather than computing direct influences on the fly

derivatives_switches::Union{DerivativesSwitch, Tuple{<:DerivativesSwitch,...}}: switch determining which of scalar potential, vector potential, velocity, and/or velocity gradient are to be computed for each target system

Optional Arguments

upward_pass::Bool: whether or not to form the multipole expansions from source bodies and translate them upward in the source tree

horizontal_pass::Bool: whether or not to transform multipole expansions from the source tree into local expansions in the target tree

downward_pass::Bool: whether or not to translate local expansions down to the leaf level of the target tree and evaluate them

unsort_bodies::Bool: indicates whether or not to undo the sort operation used to generate the octree for systems

FastMultipole.populate_influence_matrix! — Method

multiple source systems

FastMultipole.populate_influence_matrix! — Method

single source body, multiple target systems

FastMultipole.populate_influence_matrix! — Method

single source body, single target system

FastMultipole.populate_influence_matrix! — Method

single source system

FastMultipole.populate_influence_matrix! — Method

single source system (without specifying icolprev)

FastMultipole.reset! — Method

reset!(probes)

Zeroes all values (e.g. scalar/vector potential, velocity, and/or velocity gradient) of all probes.

Arguments

probes::ProbeSystem: a ::ProbeSystem object

FastMultipole.resort! — Method

Performs the same sort operation as the tree. (Undoes unsort! operation.)

FastMultipole.unsort! — Method

Undoes the sort operation performed by the tree.